requirements.txt
asgiref==3.8.1
Django==5.1.3
django-redis==5.4.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
redis==5.2.0
sqlparse==0.5.1
tzdata==2024.2
